Move decoders into separate folder
